Updated filters
Changed the way filters should be applied on an initialized instance of
brickfolio to prevent having to reinit every time to apply a new filter.
It now simply applies the filter and triggers a layout. To call the new
filter method use .brickfolio('filter', '.filter-selector'). Part of
this update was to move all the internal methods and properties to a
private namespace so as not to expose them publicly.