Fixed animation check
The check for the animation was not taking into account the mixed-delay
class when checking if any animation was being passed in at all. As the
mixed-delay class only works when coupled with an animation it should
not be counted as an animation itself and trigger the bf-animated class
being appended.