Animation demos update
Changed the pages that use the animation dropdown to have the load delay
ticked by default as changing the animation without would cause some off
the animations to look a little weird as they transitioned from the
previous animation. This should not be an issue in actual use as you
wouldn't be changing the animation on the fly as in the demos.